Move intro slideshow content to component
