+++ src/hubbleds/components/__init__.py
@@ -1,3 +1,4 @@
+from .intro_slideshow.intro_slideshow import IntroSlideshow
 from .data_table.data_table import DataTable
 from .spectrum_viewer.spectrum_viewer import SpectrumViewer
 from .spectrum_slideshow.spectrum_slideshow import SpectrumSlideshow
